Add files via upload
API stands for Application Programming Interface. On a basic level, it allows a user to "talk" to another server and request information. The API receives the request and responds back with the information.

In this example, we do not need an API key (a method of authentication) in order to request data. So think of this method as akin to web scraping, but from the back end.

We will call:
import requests    # similar to urllib, this library allows a computer to ping a website, and
import json           # library to handle JSoN formatted data